Don't bother deleting points which don't exist before importing a trace - this avoids locking the table in the common case where there are no points to delete.

+++ app/models/trace.rb
@@ -175,7 +175,12 @@ class Trace < ActiveRecord::Base
       f_lon = 0
       first = true
 
-      Tracepoint.delete_all(['gpx_id = ?', self.id])
+      # If there are any existing points for this trace then delete
+      # them - we check for existing points first to avoid locking
+      # the table in the common case where there aren't any.
+      if Tracepoint.exists?(['gpx_id = ?', self.id])
+        Tracepoint.delete_all(['gpx_id = ?', self.id])
+      end
 
       gpx.points do |point|
         if first
